scripts/requirements.txt
autopep8
boto3
pyyaml>=5.3
boto3
numpy
pycodestyle
pylint
colorama
